+++ CHANGELOG.md
@@ -6,6 +6,21 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.0-beta.9] - 2021-07-19
+
+### Added
+
+- "Show all unread topics" and "Mark all topics as read" buttons above the category
+  list in forum index
+- "Mark all topics as read" in unread topics view above and below board list
+- More logging
+- Access check when someone tries to reply to a topic they lost access to while viewing
+
+### Removed
+
+- Arbitrary `user_has_access` function from `Message` model
+
+
 ## [0.1.0-beta.8] - 2021-07-18
 
 ### Added